+++ HISTORY.rst
@@ -9,6 +9,7 @@ This version allows users to calculate normalization scaling factors,
 
 Added
 *******
+* Added Pipelines for the FASTQ module (SingleEndPipeline and PairedEndPipeline), allowing you to apply a series of functions (adapter trimming, alignment, quantification, etc) to a batch of sequence/alignment files.
 * Added new parameter to normalization functions 'return_scaling_factors', that allows you to access the scaling factors calculated by RNAlysis.
 * Added parameter 'gzip_output' to CutAdapt adapter trimming (fastq.trim_adapters_single_end and fastq.trim_adapters_paired_end), allowing users to determine whether or not the output files will be gzipped.